Use django-maploom 1.2.1
this is only because Pypi doesn't allow for pushing updates to a released package, so a minor version is added. it is still the 1.2 release of maploom but there was an update that had to be made to include the README.md. this means geoshape-geonode will have to be released as well with a minor version to accommodate the fix.